Further test refactoring and cleanup.
