+++ Tests/govuk_ios/govuk_ios_snapshot_tests/Specs/ViewControllers/RecentActivityListViewControllerSnapshotTests.swift
@@ -1,9 +1,9 @@
 import Foundation
 import XCTest
 import CoreData
-import GOVKit
 import RecentActivity
 
+@testable import GOVKitTestUtilities
 @testable import govuk_ios
 
 @MainActor
