+++ Tests/govuk_ios/govuk_ios_snapshot_tests/Specs/ViewControllers/SearchViewControllerSnapshotTests.swift
@@ -2,6 +2,7 @@ import Foundation
 import XCTest
 import UIKit
 
+@testable import GOVKitTestUtilities
 @testable import govuk_ios
 
 @MainActor
