+++ Tests/govuk_ios/govuk_ios_snapshot_tests/Specs/ViewControllers/SettingsViewControllerSnapshotTests.swift
@@ -3,6 +3,7 @@ import XCTest
 import UIKit
 import GOVKit
 
+@testable import GOVKitTestUtilities
 @testable import govuk_ios
 
 @MainActor
