+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/Builders/ViewControllerBuilderTests.swift
@@ -6,6 +6,7 @@ import Factory
 import GOVKit
 import RecentActivity
 
+@testable import GOVKitTestUtilities
 @testable import govuk_ios
 
 @MainActor
