+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/Extensions/UIKit/UIApplication+URLOpenerTests.swift
@@ -2,6 +2,7 @@ import Foundation
 import UIKit
 import Testing
 
+@testable import GOVKitTestUtilities
 @testable import govuk_ios
 
 @Suite
