+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/Models/ActivityItemCreateParamsTests.swift
@@ -1,7 +1,9 @@
 import Foundation
 import Testing
+import RecentActivity
 
 @testable import govuk_ios
+
 
 @Suite
 struct ActivityItemCreateParamsTests {
