+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/Repositories/CoreDataRepositoryTests.swift
@@ -1,6 +1,8 @@
 import Foundation
 import CoreData
 import Testing
+import GOVKitTestUtilities
+import RecentActivity
 
 @testable import govuk_ios
