+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/Services/ActivityServiceTests.swift
@@ -1,6 +1,7 @@
 import Foundation
 import Testing
 import CoreData
+import RecentActivity
 
 @testable import govuk_ios
