+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/Services/AnalyticsServiceTests.swift
@@ -2,6 +2,7 @@ import UIKit
 import Foundation
 import Testing
 import GOVKit
+import GOVKitTestUtilities
 
 @testable import govuk_ios
