+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/ViewControllers/SearchViewControllerTests.swift
@@ -4,6 +4,7 @@ import Testing
 
 import Factory
 
+@testable import GOVKitTestUtilities
 @testable import govuk_ios
 
 @MainActor
