+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/ViewModels/AppForcedUpdateContainerViewModelTests.swift
@@ -1,6 +1,7 @@
 import Foundation
 import Testing
 import GOVKit
+@testable import GOVKitTestUtilities
 @testable import govuk_ios
 
 @Suite
