+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/ViewModels/SettingsViewModelTests.swift
@@ -3,6 +3,7 @@ import Testing
 
 @testable import govuk_ios
 @testable import GOVKit
+@testable import GOVKitTestUtilities
 
 @Suite
 struct SettingsViewModelTests {
