+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/ViewModels/Topics/StepByStepsViewModelTests.swift
@@ -1,5 +1,6 @@
 import Testing
 
+@testable import GOVKitTestUtilities
 @testable import govuk_ios
 
 struct StepByStepsViewModelTests {
