+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/ViewModels/Topics/TopicDetailViewModelTests.swift
@@ -3,6 +3,7 @@ import UIKit
 
 @testable import govuk_ios
 @testable import GOVKit
+@testable import GOVKitTestUtilities
 
 struct TopicDetailViewModelTests {
